added cache

+++ src/main/java/com/github/rsallar/cineapi/service/CinemaLoadingService.java
@@ -1,8 +1,11 @@
 package com.github.rsallar.cineapi.service;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.github.rsallar.cineapi.model.Cinema;
 import com.github.rsallar.cineapi.repository.CinemaRepository;
 import com.github.rsallar.cineapi.scraping.Avui;
 
@@ -17,13 +20,15 @@ public class CinemaLoadingService {
 	@Autowired
 	CinemaRepository cinemaRepository;
 
+	
 	public void load() {
-
-		avui.getInfo().ifPresent(o -> {
+		List<Cinema> cinemas = avui.getInfo();
+			
+		if(!cinemas.isEmpty()) {
 			cinemaRepository.deleteAll();
-			cinemaRepository.save(o);
+			cinemaRepository.save(cinemas);
 			log.info("cinema info saved");
-		});
+		};
 
 	}
